requirements.txt
langchain
tiktoken
chromadb
chainlit
openai
python-dotenv
